fixing file folder
