+++ src/App.js
@@ -1,0 +1,13 @@
+import './App.css';
+
+function App() {
+
+
+  return (
+    <div className="App">
+     <h1>Main Page</h1>
+    </div>
+  );
+}
+
+export default App;
